Add "reviews" badge to post sidebar (like "flags" badge) for quick access.
Why implement this? Well, I use it all the time, so I want it
to be convenient. In particular, I use it to check up on the
actions of the reviewers for posts that I'm deleting because
they are obvious garbage. Far too often, I find reviewers in
need of some counseling and/or a ban.

There are obviously other ways to get to this, but the whole point
of this user-script is to make things easier to access, so that's
what this does.

The background color is pulled from the "hotbg" style, which
not only differentiates it from the flag count's "supernovabg"
style, but also matches the color traditionally used for
pending reviews in the classic top bar.